==> notebook/1_Select_City.ipynb <==
### Librerias 

import pymongo
from pymongo import MongoClient

import pandas as pd
from pandas import DataFrame
from pandas.io.json import json_normalize

### Conexion a Mongo 

client = MongoClient()

companies = client.ironhack.ironCollection


### Seleccion de empresas 

Aplicamos los criterios de busqueda de: 
- localizacion en territorio español
- haber cerrado rondas de financiacion por valor de 1.000.000€
- empresa fundada en el siglo XXI

filt =   {'$and':[
        {'offices.country_code': {'$eq': "ESP"}},
        {'name': {'$ne': None}},
        {'founded_year':{'$gte':2001}},
        {'offices.latitude': {'$ne': None}},
        {'offices.longitude': {'$ne': None}},
        {"funding_rounds.raised_amount":{"$gt":1_000_000}},
        {'number_of_employees':{'$gte': 87}}
                
        ]}
project =  {'_id': 0, 'name' : 1,'offices.city':1, 'category_code' : 1, 'number_of_employees' : 1,
     'founded_year' : 1, 'offices.latitude' : 1, 'offices.longitude' : 1 }
result = companies.find(filt, project)
result = list(result)
result

Como indican los resultados de la base de datos a la que tenemos acceso deberiamos instalar nuestras nuevas oficinas en la ciudad de madrid.

data = json_normalize(data =result, record_path = 'offices', meta = ['name', 'category_code', 'number_of_employees', 'founded_year'])
data

### Exportacion de datos 

data.to_csv("../data/companies.csv", index=False)
