==> notebook/2_api.ipynb <==
## Librerias

import pymongo
from pymongo import MongoClient
import geopandas as gpd
import pandas as pd
from pandas import DataFrame
from dotenv import load_dotenv
import os
import requests
import json

from functools import reduce
import operator

load_dotenv()

import sys
sys.path.append('../')
import src.Fapis as fap

## Importacion de datos

oficina = pd.read_csv('../data/companies.csv',encoding = 'ISO - 8859-1')


dfoficinas = pd.DataFrame(oficina)
dfoficinas

# comprobacion de geoenterprise
company_location = fap.geoenterprise(dfoficinas)
company_location

## LLamadas a API

### tuenti

la = company_location['Tuenti Technologies']['coordinates'][0]
lg = company_location['Tuenti Technologies']['coordinates'][1]

###### Starbucks

tok1 = os.getenv("tok1")
tok2 = os.getenv("tok2") 
url_query = 'https://api.foursquare.com/v2/venues/explore'

query = 'starbucks'
parametros = {'client_id':tok1,
              'client_secret':tok2,
              'v':'20180323',
              'll':f'{la},{lg}',
              'query': query,
              'limit':50
             }
# Generamos Request 
resp = requests.get(url=url_query,params=parametros)
data =json.loads(resp.text)


############### DECODING ###########
decoding_data = data.get("response")
decoded = decoding_data.get('groups')[0]
starbucks =decoded.get('items')

# creacion de DF con los datos del json
mapa_nombre =['venue','name']
m_latitud =['venue','location','lat']
m_longitud = ['venue','location','lng']

# introducimos los datos de json en DF 
lovemosclaro = []
for dic in starbucks:
    paralista ={}
    paralista['enterprise'] = 'Tuenti Technologies'
    paralista['team'] = 'Executives'
    paralista['requirements'] = 'starbucks'
    paralista['name'] = getFromDict(dic,mapa_nombre)
    paralista['latitud'] = getFromDict(dic,m_latitud)
    paralista['longitud'] = getFromDict(dic,m_longitud)
    lovemosclaro.append(paralista)


#creamos un data frame con lista de dicionarios si y solo si siempre son las mismas key y mismos values
gdf_starbucks_1 =pd.DataFrame(lovemosclaro)
gdf_starbucks_1[:3]

######  Museos

query = 'Museum'
parametros = {'client_id':tok1,
              'client_secret':tok2,
              'v':'20180323',
              'll':f'{la},{lg}',
              'query': query,
              'limit':50
             }
# Generamos Request 
resp = requests.get(url=url_query,params=parametros)
data =json.loads(resp.text)
############### DECODING ###########
decoding_data = data.get("response")
decoded = decoding_data.get('groups')[0]
museum =decoded.get('items')
# introducimos los datos de json en DF 
lovemosclaro = []
for dic in museum:
    paralista ={}
    paralista['enterprise'] = 'Tuenti Technologies'
    paralista['team'] = 'Designers'
    paralista['requirements'] = 'Museum'
    paralista['name'] = getFromDict(dic,mapa_nombre)
    paralista['latitud'] = getFromDict(dic,m_latitud)
    paralista['longitud'] = getFromDict(dic,m_longitud)
    lovemosclaro.append(paralista)
#creamos un data frame con lista de dicionarios si y solo si siempre son las mismas key y mismos values
gdf_museum_1 =pd.DataFrame(lovemosclaro)

gdf_museum_1[:3]

######  Aereopuerto

query = 'Airport Terminal'
parametros = {'client_id':tok1,
              'client_secret':tok2,
              'v':'20180323',
              'll':f'{la},{lg}',
              'query': query,
              'limit':50
             }
# Generamos Request 
resp = requests.get(url=url_query,params=parametros)
data =json.loads(resp.text)
############### DECODING ###########
decoding_data = data.get("response")
decoded = decoding_data.get('groups')[0]
airport =decoded.get('items')
# introducimos los datos de json en DF 
lovemosclaro = []
for dic in airport:
    paralista ={}
    paralista['enterprise'] = 'Tuenti Technologies'
    paralista['team'] = 'Account'
    paralista['requirements'] = 'Airport Terminal'
    paralista['name'] = getFromDict(dic,mapa_nombre)
    paralista['latitud'] = getFromDict(dic,m_latitud)
    paralista['longitud'] = getFromDict(dic,m_longitud)
    lovemosclaro.append(paralista)
#creamos un data frame con lista de dicionarios si y solo si siempre son las mismas key y mismos values
gdf_airport_1 =pd.DataFrame(lovemosclaro)

dfFinal_Tuenti = pd.concat([gdf_starbucks_1 , gdf_museum_1, gdf_airport_1])

dfFinal_Tuenti[:3]

### Fon

# introducimos las coordenadas de las empresas PARA FON 
laf = company_location['Tuenti Technologies']['coordinates'][0]
lgf = company_location['Tuenti Technologies']['coordinates'][1]


###### Starbucks

query = 'starbucks'
parametros = {'client_id':tok1,
              'client_secret':tok2,
              'v':'20180323',
              'll':f'{laf},{lgf}',
              'query': query,
              'limit':50
             }
# Generamos Request 
resp = requests.get(url=url_query,params=parametros)
data =json.loads(resp.text)
############### DECODING ###########
decoding_data = data.get("response")
decoded = decoding_data.get('groups')[0]
starbucks =decoded.get('items')
# introducimos los datos de json en DF 
lovemosclaro = []
for dic in starbucks:
    paralista ={}
    paralista['enterprise'] = 'Fon'
    paralista['team'] = 'Executives'
    paralista['requirements'] = 'starbucks'
    paralista['name'] = getFromDict(dic,mapa_nombre)
    paralista['latitud'] = getFromDict(dic,m_latitud)
    paralista['longitud'] = getFromDict(dic,m_longitud)
    lovemosclaro.append(paralista)
#creamos un data frame con lista de dicionarios si y solo si siempre son las mismas key y mismos values
gdf_starbucks_2 =pd.DataFrame(lovemosclaro)

###### Museos

query = 'Museum'
parametros = {'client_id':tok1,
              'client_secret':tok2,
              'v':'20180323',
              'll':f'{laf},{lgf}',
              'query': query,
              'limit':50
             }
# Generamos Request 
resp = requests.get(url=url_query,params=parametros)
data =json.loads(resp.text)
############### DECODING ###########
decoding_data = data.get("response")
decoded = decoding_data.get('groups')[0]
museum =decoded.get('items')
# introducimos los datos de json en DF 
lovemosclaro = []
for dic in museum:
    paralista ={}
    paralista['enterprise'] = 'Fon'
    paralista['team'] = 'Designers'
    paralista['requirements'] = 'Museum'
    paralista['name'] = getFromDict(dic,mapa_nombre)
    paralista['latitud'] = getFromDict(dic,m_latitud)
    paralista['longitud'] = getFromDict(dic,m_longitud)
    lovemosclaro.append(paralista)
#creamos un data frame con lista de dicionarios si y solo si siempre son las mismas key y mismos values
gdf_museum_2 =pd.DataFrame(lovemosclaro)

###### Aereopuerto

query = 'Airport Terminal'
parametros = {'client_id':tok1,
              'client_secret':tok2,
              'v':'20180323',
              'll':f'{laf},{lgf}',
              'query': query,
              'limit':50
             }
# Generamos Request 
resp = requests.get(url=url_query,params=parametros)
data =json.loads(resp.text)
############### DECODING ###########
decoding_data = data.get("response")
decoded = decoding_data.get('groups')[0]
airport =decoded.get('items')
# introducimos los datos de json en DF 
lovemosclaro = []
for dic in airport:
    paralista ={}
    paralista['enterprise'] = 'Fon'
    paralista['team'] = 'Account'
    paralista['requirements'] = 'Airport Terminal'
    paralista['name'] = getFromDict(dic,mapa_nombre)
    paralista['latitud'] = getFromDict(dic,m_latitud)
    paralista['longitud'] = getFromDict(dic,m_longitud)
    lovemosclaro.append(paralista)
#creamos un data frame con lista de dicionarios si y solo si siempre son las mismas key y mismos values
gdf_airport_2 =pd.DataFrame(lovemosclaro)

dfFinal_Fon= pd.concat([gdf_starbucks_2 , gdf_museum_2, gdf_airport_2])
dfFinal_Fon[:3] 

dfFinal = pd.concat([dfFinal_Fon , dfFinal_Tuenti])
dfFinal[:3]

## Exportacion de datos

dfFinal.to_csv("../data/final.csv", index = False)

dfFinal_Fon.to_csv("../data/Fon.csv", index = False)

dfFinal_Tuenti.to_csv("../data/Tuenti.csv", index = False)
