==> notebook/3_geospatial .ipynb <==
## Librerias 

from pymongo import MongoClient
import pandas as pd
import geopandas as gdp
import shapely
import json
from bson.json_util import dumps
from pandas import DataFrame

## Importacion de datos 

dft = pd.read_csv('../data/Tuenti.csv')
dft

dffon = pd.read_csv('../data/Fon.csv')
dffon

## GeoData

Tratamiento de las latitudes y las longitudes de las bases de datos iniciales para su  manipulacion en mongo, obteniendo asi las distancias relativas a las distintas localizaciones respecto a los candidatos Tuenti y Fon.

# generacion de bbdd y collecion 
client = MongoClient()
db = client.geoloca # creamos bbdd
collection = db.tuenti # creamos collecion para tuenti
collection = db.fon # creamos collecion para fon

#####  Tuenti

# generamos geometry que es el geojson

dft = gdp.GeoDataFrame(dft, geometry= gdp.points_from_xy(dft.longitud, dft.latitud ))
dft.columns=['enterprise', 'team','requirements','name','latitud','longitud','geometry' ]
    
#Lo aplicamos a toda la columna
dft['geometry']= dft['geometry'].apply(lambda x:shapely.geometry.mapping(x))

# creamos index pasando una clave para cada objeto geoespacial y pasamos 2dspher para que lo entienda mongo 
collection.create_index([("geometry", "2dsphere")]) 

datat = dft.to_dict(orient='records')
datat

collection.insert_many(datat)

calculamos distancias haciendo una agragacion mediante $geoNear 

  ########## # para tuenti 40.415824	-3.696277 # ###########  	    
fillter =  [{ "$geoNear": {'near': [-3.696277, 40.415824],
                 'distanceField': 'distance',
                 'maxDistance': 1000,
                 'distanceMultiplier': 6371, 
                 'spherical'  : True}}]
distance = collection.aggregate(fillter)
tuenti_result = json.loads(dumps(distance))

distance_tuenti = pd.DataFrame(list(tuenti_result))

#####  Fon

dffon = gdp.GeoDataFrame(dffon, geometry= gdp.points_from_xy(dffon.longitud, dffon.latitud ))
dffon.columns=['enterprise', 'team','requirements','name','latitud','longitud','geometry' ]    
dffon['geometry']= dffon['geometry'].apply(lambda x:shapely.geometry.mapping(x))

collection.create_index([("geometry", "2dsphere")]) 

dataf = dffon.to_dict(orient='records')

collection.insert_many(dataf)

  ########## #  Fon 40.530724	-3.632687 # ###########  	   
fillter =  [{ "$geoNear": {'near': [-3.632687, 40.530724],
                 'distanceField': 'distance',
                 'maxDistance': 7000,
                 'distanceMultiplier': 6371, 
                 'spherical'  : True}}]
distance = collection.aggregate(fillter)
fon_result = json.loads(dumps(distance))

distance_fon = pd.DataFrame(list(fon_result))

Una vez obtenidas las distancias podemos comparar los distintos requerimiento de los  departamentos de la empresa.  

recap = pd.concat([distance_tuenti,distance_fon])
recap[:3]

final = recap.groupby(['enterprise','requirements']).mean()
final

top = final.sort_values('distance')
top

Decidimo emplazar las nuevas oficinas en los alrededores de "Tuenti Technologies" ya que la mayoria de las distancias medias de los servicios y localizaciones son alli menores.
